automate/requirements.txt
groq
requests
python-dotenv
apscheduler
fal-client
